Cole/switch to ef

+++ NhlGameDatabase/Tables/PredictedGame.sql
@@ -4,9 +4,9 @@
     homeTeamName VARCHAR(MAX) NOT NULL,
     awayTeamName VARCHAR(MAX) NOT NULL,
     gameDate Datetime NOT NULL,
-    vegasHomeOdds DECIMAL(8, 5) NOT NULL DEFAULT 0,
-    vegasAwayOdds DECIMAL(8, 5) NOT NULL DEFAULT 0,
-    modelHomeOdds DECIMAL(8, 5) NOT NULL DEFAULT 0,
-    modelAwayOdds DECIMAL(8, 5) NOT NULL DEFAULT 0,
+    vegasHomeOdds FLOAT NOT NULL DEFAULT 0,
+    vegasAwayOdds FLOAT NOT NULL DEFAULT 0,
+    modelHomeOdds FLOAT NOT NULL DEFAULT 0,
+    modelAwayOdds FLOAT NOT NULL DEFAULT 0,
     PRIMARY KEY(id),
 );
